fix: directories refactored
