+++ shcharbunou/lesson10-11/music-store/src/main/java/by/shcharbunou/musicstore/servlet/add/AddArtistServlet.java
@@ -1,4 +1,4 @@
-package by.shcharbunou.musicstore.servlet;
+package by.shcharbunou.musicstore.servlet.add;
 
 import by.shcharbunou.musicstore.dao.ArtistDao;
 import by.shcharbunou.musicstore.dao.impl.ArtistDaoImpl;
