+++ shcharbunou/lesson10-11/music-store/src/main/java/by/shcharbunou/musicstore/servlet/update/UpdateArtistServlet.java
@@ -1,4 +1,4 @@
-package by.shcharbunou.musicstore.servlet;
+package by.shcharbunou.musicstore.servlet.update;
 
 import by.shcharbunou.musicstore.dao.ArtistDao;
 import by.shcharbunou.musicstore.dao.ArtistSongDao;
