+++ shcharbunou/lesson10-11/music-store/src/main/java/by/shcharbunou/musicstore/servlet/update/UpdateSongServlet.java
@@ -1,0 +1,20 @@
+package by.shcharbunou.musicstore.servlet.update;
+
+import jakarta.servlet.*;
+import jakarta.servlet.http.*;
+import jakarta.servlet.annotation.*;
+
+import java.io.IOException;
+
+@WebServlet(name = "UpdateSongServlet", value = "/UpdateSongServlet")
+public class UpdateSongServlet extends HttpServlet {
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        doPut(request, response);
+    }
+
+    @Override
+    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+    }
+}
