+++ shcharbunou/lesson10-11/music-store/src/main/webapp/jsp-content/delete-jsp/delete-artist.jsp
@@ -2,12 +2,12 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
-    <title>Delete artist</title>
+    <title>Delete Artist</title>
 </head>
 <body>
 <form action="<c:url value="/delete-artist"/>" method="post">
     <fieldset>
-        <legend>Delete artist</legend>
+        <legend>Delete Artist</legend>
         <label for="name">Name</label>
         <input name="name" id="name" type="text" required oninvalid="this.setCustomValidity('Enter artist')" oninput="setCustomValidity('')">
         <br>
