+++ shcharbunou/lesson10-11/music-store/src/main/webapp/jsp-content/delete-jsp/delete-song.jsp
@@ -2,12 +2,12 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
-    <title>Delete song</title>
+    <title>Delete Song</title>
 </head>
 <body>
 <form action="<c:url value="/delete-song"/>" method="post">
     <fieldset>
-        <legend>Delete song</legend>
+        <legend>Delete Song</legend>
         <label for="artist">Artist</label>
         <input name="artist" id="artist" type="text" required oninvalid="this.setCustomValidity('Enter artist')" oninput="setCustomValidity('')">
         <br>
