+++ shcharbunou/lesson10-11/music-store/src/main/webapp/jsp-content/update-jsp/update-artist.jsp
@@ -2,16 +2,16 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
-    <title>Update artist</title>
+    <title>Update Artist</title>
 </head>
 <body>
 <form action="<c:url value="/update-artist"/>" method="post">
     <fieldset>
-        <legend>Update artist</legend>
+        <legend>Update Artist</legend>
         <label for="old_artist">Old Artist</label>
         <input name="old_artist" id="old_artist" type="text" required oninvalid="this.setCustomValidity('Enter old artist')" oninput="setCustomValidity('')">
         <br>
-        <label for="new_artist">New Artist</label>
+        <label for="new_artist">Updated Artist</label>
         <input name="new_artist" id="new_artist" type="text" required oninvalid="this.setCustomValidity('Enter new artist')" oninput="setCustomValidity('')">
         <br>
         <label for="new_country">Country</label>
